pyproject.toml
[tool.poetry]
name = "intro-pypsa"
version = "0.1.0"
description = ""
authors = [
    "Carlo Federico Vescovo <cfv@ieee.org>",
    "Giovanni Zaccaria <s295335@studenti.polito.it>",
    "Matteo Savina <matteo.savina@polito.it>"
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">= 3.10, < 3.11"
numpy = "^1.26.1"
pypsa = { git = "https://github.com/pso-energy/PyPSA.git" }
gurobipy = "^11.0.2"
python-dotenv = "^1.0.1"
zipfile36 = "^0.1.3"
streamerate = "^1.1.0"
psutil = "^5.9.8"
requests = "^2.32.3"
pytest = "^8.2.2"
seaborn = "^0.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
